evals/voice_evals/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-evals"
version = "0.1.0"
description = "Comprehensive voice AI evaluation toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
]

[project.optional-dependencies]
asr = [
    "openai-whisper>=20231117",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "jiwer>=3.0.0",
    "tiktoken>=0.5.0",
]
tts = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]
agent = [
    "openai>=1.0.0",
]
diarization = [
    "pyannote.audio>=3.1.0",
]
full = [
    "voice-evals[asr,tts,agent,diarization]",
    "funasr>=1.0.0",
]

[project.scripts]
voice-evals = "voice_evals.cli:main"

[tool.setuptools.packages.find]
include = ["voice_evals*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: requires ML model downloads (deselect with '-m \"not slow\"')",
]
